001 cleanup and minor refactoring
