+++ src/main/kotlin/com/guidovezzoni/gradle/smartproperties/model/Type.kt
