+++ src/main/kotlin/com/guidovezzoni/gradle/smartproperties/model/VariantInfo.kt
